+++ packages/enterprise/src/lib/enterprise.spec.ts
@@ -1,7 +1,0 @@
-import { enterprise } from './enterprise';
-
-describe('enterprise', () => {
-  it('should work', () => {
-    expect(enterprise()).toEqual('enterprise');
-  });
-});