+++ packages/enterprise/src/lib/enterprise.ts
@@ -1,3 +1,0 @@
-export function enterprise(): string {
-  return 'enterprise';
-}