+++ scripts/ci-publish.sh
@@ -48,7 +48,6 @@ PACKAGES=(
     "events"
     "ng-forms"
     "realtime"
-    "enterprise"
     "guardrails"
     "schema"
 )